Discovered issue with Timer/RTC/IO board
The original design of the Timer/RTC/IO board did not cater for a situation where multiple Z80 peripherals may exist in the system.

Consequently, when M1/ and IORQ/ are both asserted by the CPU to read a vector from the interrupting peripheral, and if that peripheral happens to be the SIO, the Timer/RTC/IO board will configure its data bus buffer for output towards the CPU. If the SIO is responding with its vector, this causes contention on that portion of the bus, resulting in interupts to random vectors.

Staged some logic for a subsequent modification (gates U14B and U13A) which would prevent the Timer/RTC/IO board from configuring its data bus buffer outwards if the IEI/ signal is asserted. However, this logic isnt latched in any way, so it could be interrupted half way through if IEI/ is asserted while outputting a vector... Perhaps more thought needed.