Interrupt daisy chain modification
To overcome the issue identified earlier, it seems better to change the interrupt daisy chain to make the CTC the higher priority device. This also simplifies the extra required circuitry to only a single OR gate (U13A).

Now, only when the IEO/ signal of the CTC is asserted will the board drive the data bus to the CPU.

Also added in a previous revision is changes to the beeper circuit. Will now be driven by a 555 timer, gated by the output register.