Reverse order of semaphore and mutex check
This was done in line with changes to ctc_isr.s which only increments an apps semaphore if that app owns the foreground mutex.

This fixes a small glitch where the semaphore was continually incremented, and if you happened to exit an app when the next apps semaphore was 0, you would return immediately to the app you just exited.